requirements.txt
fastapi
uvicorn[standard]
gunicorn
twilio
openai
python-dotenv
openpyxl
pydantic
python-multipart
requests
